+++ wzdx/tools/gcp_tools.py
@@ -43,22 +43,3 @@ def formatMessage(message) -> str:
     else:
         msg = str(message)
     return msg
-
-
-def get_wzdx_schema(schema_file_name):
-    try:
-        schema_string = open(schema_file_name, 'r').read()
-    except FileNotFoundError as e:
-        raise RuntimeError('invalid schema: file does not exist')
-    if not schema_string:
-        raise RuntimeError('invalid schema: empty schema')
-    try:
-        schema_obj = json.loads(schema_string)
-        jsonschema.validate('', schema_obj)
-    except json.decoder.JSONDecodeError as e:
-        raise RuntimeError('invalid schema: not valid json') from e
-    except jsonschema.SchemaError as e:
-        raise RuntimeError('invalid schema: schema failed validation') from e
-    except jsonschema.ValidationError:
-        return schema_obj
-    return schema_obj
